# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: NTT DevSecOps Assessment Internal API
  version: 1.0.0-phase6
  description: API interna versionada para catálogo SAMM, assessments, respuestas, evidencias, resultados, recomendaciones y reportes.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
  /health:
    get:
      security: []
      summary: Estado de salud de la aplicación
      responses:
        '200':
          description: Aplicación disponible
  /questionnaire-versions/:
    get:
      summary: Lista versiones del cuestionario
      responses:
        '200': {description: Versiones visibles}
        '401': {description: Sesión requerida}
  /questions/:
    get:
      summary: Lista preguntas maestras activas
      parameters:
        - in: query
          name: q
          schema: {type: string}
      responses:
        '200': {description: Catálogo de preguntas}
  /questions/{question_id}/:
    get:
      summary: Detalle de una pregunta maestra
      parameters:
        - $ref: '#/components/parameters/QuestionId'
      responses:
        '200': {description: Pregunta y alternativas}
        '404': {description: Pregunta no encontrada}
  /assessments/:
    get:
      summary: Lista assessments autorizados para el usuario
      responses:
        '200': {description: Assessments y avance}
  /assessments/{assessment_id}/:
    get:
      summary: Detalle de un assessment autorizado
      parameters:
        - $ref: '#/components/parameters/AssessmentId'
      responses:
        '200': {description: Assessment, asignaciones y progreso}
        '403': {description: Usuario no autorizado}
  /assessments/{assessment_id}/questions/:
    get:
      summary: Preguntas instanciadas del assessment
      parameters:
        - $ref: '#/components/parameters/AssessmentId'
      responses:
        '200': {description: Preguntas, snapshots y respuestas}
        '403': {description: Usuario no autorizado}
  /responses/{assessment_question_id}/:
    put:
      summary: Guarda borrador, respuesta o envío a revisión
      parameters:
        - in: path
          name: assessment_question_id
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponseWrite'
      responses:
        '200': {description: Respuesta guardada}
        '403': {description: El usuario no es respondedor asignado}
        '409': {description: Transición o datos inválidos}
  /evidences/{evidence_id}/:
    get:
      summary: Metadatos de una evidencia autorizada
      parameters:
        - in: path
          name: evidence_id
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200': {description: Metadatos, hash y estado de validación}
        '403': {description: Acceso denegado}
  /results/{assessment_id}/:
    get:
      summary: Resultado calculado o publicado del assessment
      description: Administradores y revisores autorizados pueden previsualizar. El respondedor requiere resultados publicados.
      parameters:
        - $ref: '#/components/parameters/AssessmentId'
        - in: query
          name: include_questions
          schema: {type: boolean, default: false}
      responses:
        '200':
          description: Resultado general, resumen y dimensiones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentResult'
        '403': {description: Resultados no publicados o acceso denegado}
        '404': {description: Assessment no encontrado}
  /recommendations/:
    get:
      summary: Lista recomendaciones de un assessment autorizado
      parameters:
        - in: query
          name: assessment_id
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Recomendaciones activas
        '400': {description: Falta assessment_id}
        '403': {description: Acceso denegado}
  /reports/{assessment_id}/:
    get:
      summary: Obtiene los enlaces autorizados para reportes HTML, Excel y PDF
      parameters:
        - $ref: '#/components/parameters/AssessmentId'
      responses:
        '200': {description: Formatos y rutas de descarga disponibles}
        '403': {description: Resultados no publicados o acceso denegado}
        '404': {description: Assessment no encontrado}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: samm_session
  parameters:
    AssessmentId:
      in: path
      name: assessment_id
      required: true
      schema: {type: string, format: uuid}
    QuestionId:
      in: path
      name: question_id
      required: true
      schema: {type: string, format: uuid}
  schemas:
    ResponseWrite:
      type: object
      properties:
        selected_option_code: {type: [string, 'null']}
        respondent_comment: {type: [string, 'null'], maxLength: 10000}
        is_not_applicable: {type: boolean, default: false}
        not_applicable_justification: {type: [string, 'null'], maxLength: 5000}
        intent:
          type: string
          enum: [autosave, save, submit]
          default: save
    AssessmentResult:
      type: object
      required: [source, calculation_version, overall_score, progress_percent, status_summary, dimensions]
      properties:
        source:
          type: string
          enum: [declared, reviewed, approved]
        calculation_version: {type: string}
        input_hash: {type: string, minLength: 64, maxLength: 64}
        overall_score: {type: number, minimum: 0, maximum: 3}
        progress_percent: {type: number, minimum: 0, maximum: 100}
        target_score: {type: [number, 'null'], minimum: 0, maximum: 3}
        gap: {type: [number, 'null'], minimum: 0, maximum: 3}
        status_summary: {type: object}
        metadata: {type: object}
        dimensions:
          type: array
          items:
            $ref: '#/components/schemas/ScoreDimension'
        snapshot:
          type: [object, 'null']
    ScoreDimension:
      type: object
      properties:
        dimension_type:
          type: string
          enum: [question, level, stream, practice, function, overall]
        key: {type: string}
        name: {type: string}
        parent_key: {type: [string, 'null']}
        score: {type: number}
        max_score: {type: number}
        normalized_percent: {type: number}
        target_score: {type: [number, 'null']}
        gap: {type: [number, 'null']}
        applicable_count: {type: integer}
        not_applicable_count: {type: integer}
        pending_count: {type: integer}
        metadata: {type: object}